refactor(checker): split diagnostic source collection display
